renewal frontend with npm and webpack
* remove asset_pipeline and rails-assets
* add npm and webpack
