+++ app.json
@@ -8,5 +8,13 @@
   "addons":[
     "memcachier",
     "mongolab"
+  ],
+  "buildpacks": [
+    {
+      "url": "https://github.com/heroku/heroku-buildpack-ruby"
+    },
+    {
+      "url": "https://github.com/heroku/heroku-buildpack-nodejs"
+    }
   ]
 }
