--- assets/css/app.css
@@ -1,3 +1,0 @@
-/*
- *= require_tree .
- */
