--- assets/js/app.js
@@ -1,3 +1,0 @@
-//= require react
-//= require jquery
-//= require_tree .
