+++ bin/run.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+npm run production
+bundle exec thin start -p $PORT
