+++ package.json
@@ -1,0 +1,32 @@
+{
+  "name": "bookscan-premium-optimizer",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "production": "webpack -p",
+    "develop": "webpack -d",
+    "watch": "webpack -d --watch",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "TADA, Tadashi",
+  "license": "GPL",
+  "engine": {
+	  "node": "^6.1.0",
+	  "npm": "^3.8.6"
+  },
+  "dependencies": {
+    "jquery": "^2.2.3",
+    "react": "^15.0.2",
+    "react-dom": "^15.0.2",
+    "babel-core": "^6.8.0",
+    "babel-loader": "^6.2.4",
+    "babel-preset-es2015": "^6.6.0",
+    "babel-preset-react": "^6.5.0",
+    "css-loader": "^0.23.1",
+    "style-loader": "^0.13.1",
+    "webpack": "^1.13.0"
+  },
+  "devDependencies": {
+  }
+}
