+++ views/index.haml
@@ -3,7 +3,6 @@
 	%head
 		%meta{charset: 'UTF-8'}
 		%title BOOKSCAN Premium Optimizer
-		!= stylesheet_tag 'app'
 	%body
 		%h1.top BOOKSCAN Premium Optimizer
 		#main
@@ -23,4 +22,4 @@
 			%p.desc
 				最近利用したリスト:
 			#recent-list
-		!= javascript_tag 'app'
+		%script{src: "/js/app.js"}
