+++ views/list.haml
@@ -3,7 +3,6 @@
 	%head
 		%meta{charset: 'UTF-8'}
 		%title BOOKSCAN Premium Optimizer
-		!= stylesheet_tag 'app'
 	%body{id: booklist_id}
 		%h1 BOOKSCAN Premium Optimizer
 		#contents
@@ -16,4 +15,4 @@
 			リストから外したい本があれば、題名の左にある[×]を押して下さい。
 		%p.desc
 			このページをブックマークしておくと、あとで再利用できます。
-		!= javascript_tag 'app'
+		%script{src: "/js/app.js"}
